LayoutTests/inspector/debugger/function-location.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>

<script>  function firstLineFunction()

{
}

   function notFirstLineFunction()

{
}

function test()
{
    InspectorTest.runDebuggerTestSuite([
        function testGetFirstLineFunctionLocation(next)
        {
            InspectorTest.evaluateInPage("firstLineFunction", didEvaluate);

            function didEvaluate(remote)
            {
                InspectorTest.addResult("firstLineFunction type = " + remote.type);
                DebuggerAgent.getFunctionLocation(remote.objectId, didGetLocation);
            }
            function didGetLocation(error, response)
            {
                InspectorTest.addResult("didGetLocation: ");
                InspectorTest.addResult("lineNumber: " + response.lineNumber);
                InspectorTest.addResult("columnNumber: " + response.columnNumber);
                InspectorTest.addResult("scriptId is valid: " + !!response.scriptId);
                next();
            }
        },
        function testGetNonFirstLineFunctionLocation(next)
        {
            InspectorTest.evaluateInPage("notFirstLineFunction", didEvaluate);

            function didEvaluate(remote)
            {
                InspectorTest.addResult("nonFirstLineFunctiontype = " + remote.type);
                DebuggerAgent.getFunctionLocation(remote.objectId, didGetLocation);
            }
            function didGetLocation(error, response)
            {
                InspectorTest.addResult("didGetLocation: ");
                InspectorTest.addResult("lineNumber: " + response.lineNumber);
                InspectorTest.addResult("columnNumber: " + response.columnNumber);
                InspectorTest.addResult("scriptId is valid: " + !!response.scriptId);
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests that Debugger.getFunctionLocation command returns location.
<a href="https://bugs.webkit.org/show_bug.cgi?id=71808">Bug 71808</a>
</p>

</body>
</html>
