LayoutTests/inspector/debugger/function-location-expected.txt

Tests that Debugger.getFunctionLocation command returns location. Bug 71808

Debugger was enabled.

Running: testGetFirstLineFunctionLocation
firstLineFunction type = function
didGetLocation: 
lineNumber: 7
columnNumber: undefined
scriptId is valid: true

Running: testGetNonFirstLineFunctionLocation
nonFirstLineFunctiontype = function
didGetLocation: 
lineNumber: 12
columnNumber: undefined
scriptId is valid: true
Debugger was disabled.
